+++ SAM-6D/Pose_Estimation_Model/run_inference_custom.py
@@ -123,7 +123,7 @@ def _get_template(path, cfg, tem_index=1):
 
     rgb = load_im(rgb_path).astype(np.uint8)
     xyz = np.load(xyz_path).astype(np.float32) / 1000.0  
-    mask = load_im(mask_path).astype(int) == 255
+    mask = load_im(mask_path).astype(np.uint8) == 255
 
     bbox = get_bbox(mask)
     y1, y2, x1, x2 = bbox
